refactor(mitre): move cve packages under mitre/cve to mirror cache dirs
The package path conventionally mirrors the cache-dir hierarchy in this repo.
Now that the cache dirs / ids / subcommands for cvrf/v4/v5 are nested under
cve, relocate the packages to match:

  pkg/{fetch,extract}/mitre/{cvrf,v4,v5} -> pkg/{fetch,extract}/mitre/cve/{cvrf,v4,v5}

Package names are unchanged; only import paths/aliases are updated. Pure
internal move with no downstream impact.
