+++ pkg/cmd/extract/extract.go
@@ -67,10 +67,10 @@ import (
 	microsoftMSUC "github.com/MaineK00n/vuls-data-update/pkg/extract/microsoft/msuc"
 	microsoftWSUSSCN2 "github.com/MaineK00n/vuls-data-update/pkg/extract/microsoft/wsusscn2"
 	mitreCAPEC "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/capec"
-	mitreCVRF "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/cvrf"
+	mitreCVRF "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/cve/cvrf"
+	mitreV4 "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/cve/v4"
+	mitreV5 "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/cve/v5"
 	mitreCWE "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/cwe"
-	mitreV4 "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/v4"
-	mitreV5 "github.com/MaineK00n/vuls-data-update/pkg/extract/mitre/v5"
 	"github.com/MaineK00n/vuls-data-update/pkg/extract/msf"
 	"github.com/MaineK00n/vuls-data-update/pkg/extract/netbsd"
 	npmDB "github.com/MaineK00n/vuls-data-update/pkg/extract/npm/db"
