+++ pkg/cmd/fetch/fetch.go
@@ -96,11 +96,11 @@ import (
 	minimOSSecDB "github.com/MaineK00n/vuls-data-update/pkg/fetch/minimos/secdb"
 	mitreATTACK "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/attack"
 	mitreCAPEC "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/capec"
-	mitreCVRF "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/cvrf"
+	mitreCVRF "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/cve/cvrf"
+	mitreV4 "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/cve/v4"
+	mitreV5 "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/cve/v5"
 	mitreCWE "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/cwe"
 	mitreEMB3D "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/emb3d"
-	mitreV4 "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/v4"
-	mitreV5 "github.com/MaineK00n/vuls-data-update/pkg/fetch/mitre/v5"
 	"github.com/MaineK00n/vuls-data-update/pkg/fetch/msf"
 	ncscCSAF "github.com/MaineK00n/vuls-data-update/pkg/fetch/ncsc/csaf"
 	"github.com/MaineK00n/vuls-data-update/pkg/fetch/netbsd"
