+++ pkg/extract/mitre/cve/cvrf/testdata/fixtures/2021/CVE-2021-36622.json
