+++ pkg/extract/mitre/cve/v4/testdata/fixtures/1999/CVE-1999-0001.json
