+++ pkg/extract/mitre/cve/v4/v4.go
