+++ pkg/extract/mitre/cve/v5/testdata/fixtures/2024/CVE-2024-39565.json
