+++ pkg/extract/mitre/cve/v5/testdata/fixtures/2024/CVE-2024-40946.json
