+++ pkg/extract/mitre/cve/v5/testdata/golden/data/2024/CVE-2024-40946.json
