+++ pkg/extract/mitre/cve/v5/testdata/golden/data/2024/CVE-2024-6958.json
