+++ pkg/extract/mitre/cve/v5/testdata/golden/datasource.json
