+++ pkg/fetch/mitre/cve/cvrf/testdata/fixtures/allitems-cvrf.xml
