+++ pkg/fetch/mitre/cve/cvrf/testdata/golden/2021/CVE-2021-36622.json
