+++ pkg/fetch/mitre/cve/cvrf/types.go
