+++ pkg/fetch/mitre/cve/v4/testdata/fixtures/allitems.xml.gz
