+++ pkg/fetch/mitre/cve/v4/testdata/golden/1999/CVE-1999-0002.json
