+++ pkg/fetch/mitre/cve/v4/types.go
