+++ pkg/fetch/mitre/cve/v4/v4.go
