+++ pkg/fetch/mitre/cve/v5/testdata/fixtures/cvelistV5-main.tar.gz
