+++ pkg/fetch/mitre/cve/v5/testdata/golden/2022/CVE-2022-0002.json
