+++ pkg/fetch/mitre/cve/v5/testdata/golden/2023/CVE-2023-0001.json
