+++ pkg/fetch/mitre/cve/v5/types.go
